# faults-openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"https://localhost:4443","description":"Generated server url"}],"paths":{"/faults/{faultId}":{"get":{"tags":["fault-controller"],"operationId":"retrieveById","parameters":[{"name":"faultId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FaultResponseDto"}}}}}},"put":{"tags":["fault-controller"],"operationId":"update","parameters":[{"name":"faultId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FaultEditDto"}}},"required":true},"responses":{"404":{"description":"Not Found"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["fault-controller"],"operationId":"delete","parameters":[{"name":"faultId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/faults":{"get":{"tags":["fault-controller"],"operationId":"retrieveMany","parameters":[{"name":"isClosedIncluded","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"404":{"description":"Not Found"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FaultResponseDto"}}}}}}},"post":{"tags":["fault-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FaultEditDto"}}},"required":true},"responses":{"404":{"description":"Not Found"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/faults/{faultId}/attachments":{"post":{"tags":["fault-controller"],"operationId":"createAttachment","parameters":[{"name":"faultId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["attachment"],"type":"object","properties":{"attachment":{"type":"string","format":"binary"}}}}}},"responses":{"404":{"description":"Not Found"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/faults/{faultId}/attachments/{attachmentName}":{"get":{"tags":["fault-controller"],"operationId":"retrieveAttachmentById","parameters":[{"name":"faultId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"attachmentName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}},"delete":{"tags":["fault-controller"],"operationId":"deleteAttachment","parameters":[{"name":"faultId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"attachmentName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"FaultEditDto":{"type":"object","properties":{"description":{"type":"string"},"closed":{"type":"boolean"}}},"FaultResponseDto":{"type":"object","properties":{"uri":{"type":"string","format":"uri"},"description":{"type":"string"},"openedBy":{"type":"string"},"openedAt":{"type":"integer","format":"int64"},"closedBy":{"type":"string"},"closedAt":{"type":"integer","format":"int64"},"attachments":{"type":"array","items":{"type":"string","format":"uri"}}}}}}}